refactor(angular_forms): Move common functionality of NgForm and NgFormModel into a new AbstractForm class.
PiperOrigin-RevId: 187346765
