+++ angular_forms/lib/src/directives/ng_form.dart
@@ -1,13 +1,12 @@
 import 'dart:async';
-import 'dart:html' show Event;
 
 import 'package:meta/meta.dart';
 import 'package:angular/angular.dart';
 
 import '../model.dart' show AbstractControl, ControlGroup, Control;
 import '../validators.dart' show NG_VALIDATORS;
+import 'abstract_form.dart' show AbstractForm;
 import 'control_container.dart' show ControlContainer;
-import 'form_interface.dart' show Form;
 import 'ng_control.dart' show NgControl;
 import 'ng_control_group.dart' show NgControlGroup;
 import 'shared.dart' show setUpControl, setUpControlGroup, composeValidators;
@@ -72,33 +71,15 @@ import 'shared.dart' show setUpControl, setUpControlGroup, composeValidators;
   providers: const [
     const ExistingProvider(ControlContainer, NgForm),
   ],
-  host: const {'(submit)': 'onSubmit(\$event)'},
   exportAs: 'ngForm',
   visibility: Visibility.all,
 )
-class NgForm extends ControlContainer implements Form {
+class NgForm extends AbstractForm {
   ControlGroup form;
-  final _ngSubmit = new StreamController<ControlGroup>.broadcast(sync: true);
-  final _ngBeforeSubmit =
-      new StreamController<ControlGroup>.broadcast(sync: true);
 
   NgForm(@Optional() @Self() @Inject(NG_VALIDATORS) List<dynamic> validators) {
     form = new ControlGroup({}, null, composeValidators(validators));
   }
-
-  @Output()
-  Stream<ControlGroup> get ngSubmit => _ngSubmit.stream;
-  @Output()
-  Stream<ControlGroup> get ngBeforeSubmit => _ngBeforeSubmit.stream;
-
-  @override
-  Form get formDirective => this;
-
-  @override
-  ControlGroup get control => form;
-
-  @override
-  List<String> get path => [];
 
   Map<String, AbstractControl> get controls => form.controls;
 
@@ -112,9 +93,6 @@ class NgForm extends ControlContainer implements Form {
       ctrl.updateValueAndValidity(emitEvent: false);
     });
   }
-
-  @override
-  Control getControl(NgControl dir) => (form.find(dir.path) as Control);
 
   @override
   void removeControl(NgControl dir) {
@@ -150,21 +128,10 @@ class NgForm extends ControlContainer implements Form {
   }
 
   @override
-  ControlGroup getControlGroup(NgControlGroup dir) =>
-      (form.find(dir.path) as ControlGroup);
-
-  @override
   void updateModel(NgControl dir, dynamic value) {
     scheduleMicrotask(() {
-      Control ctrl = form.find(dir.path);
-      ctrl?.updateValue(value);
+      super.updateModel(dir, value);
     });
-  }
-
-  void onSubmit(Event event) {
-    _ngBeforeSubmit.add(form);
-    _ngSubmit.add(form);
-    event?.preventDefault();
   }
 
   @protected
